Revert "Don't update bounds on each map move (#906)"
This reverts commit b026e1f1ff068ab45ede14c48979c1959d1b03a1.
